added css and new pictures
I think this is the most fun that I've ever had coding. To be honest, I personally think Alexis would like CSS. She can work with a lot of pink, and design her own page full of her favorite things. She can also let her mind floe and not be so in the box with coding. However, that is my personal opinion and a suggestion for her.